Made CounterUnit its own class & refactored JobCounter to work with it.
